langchain/requirements.txt
langchain==0.0.352
python-dotenv==1.0.0
openai==1.3.7
chromadb==0.4.18
pypdf==3.17.1
tiktoken==0.5.1
